pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.poetry]
name = "hello"
version = "0.0.1"
description = "Custom Decision Transformer Modeling."
license = "MIT"
authors = [
    "Yoann Poupart <yoann.poupart@ens-lyon.org>",
    "Sam Johnson <sejohnson2020@outlook.com>"
]

[[tool.poetry.source]]
name = "PyPI"
priority = "default"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
torch = ">=2.0.0, !=2.0.1, !=2.1.0"
transformers = { version = "^4.3", extras = ["torch"] }
torchtyping = "^0.1.4"
python-chess = "^1.999"
jsonlines = "^4.0.0"
tensorboard = "^2.15.0"
evaluate = "^0.4.1"
spicy = "^0.16.0"
scikit-learn = "^1.3.2"
matplotlib = "^3.8.2"
torchrl = "^0.2.1"
awslambdaric = "^2.0.8"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
ipykernel = "^6.29.3"
notebook = "^7.1.1"
